requirements/base.txt
Django==1.10

# Configuration
django-environ==0.4.0
whitenoise==3.2

# Network
requests==2.10.0
django-subdomains==2.1.0

# Views
django-extra-views==0.8.0

# Parsing
lxml==3.6.1
